Separate example from library
