+++ .github/workflows/ci.yml
@@ -1,0 +1,62 @@
+name: CI
+on:
+  push:
+    branches: ['master']
+    tags: ['*']
+  pull_request:
+    branches: ['master']
+
+jobs:
+  test:
+    name: Test
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        project_dir:
+          - .
+          - examples/express
+          - examples/express-apikey-auth
+          - examples/express-jwt-auth
+          - examples/express-ts-mock
+          - examples/express-typescript
+          - examples/hapi-typescript
+          - examples/koa
+          - examples/serverless-framework
+          - examples/aws-sam
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions/setup-node@v3
+        with:
+          node-version: '12'
+      - uses: actions/setup-python@v3
+        with:
+          python-version: '3.x'
+      - run: pip install aws-sam-cli
+        if: matrix.project_dir == 'examples/aws-sam'
+      - run: npm ci
+        working-directory: ${{ matrix.project_dir }}
+      - run: npm link ../../
+        if: matrix.project_dir != '.' && matrix.project_dir != 'examples/aws-sam'
+        working-directory: ${{ matrix.project_dir }}
+      - run: npm run lint
+        if: matrix.project_dir == '.'
+        working-directory: ${{ matrix.project_dir }}
+      - run: npm test
+        working-directory: ${{ matrix.project_dir }}
+
+  publish:
+    name: Publish
+    runs-on: ubuntu-latest
+    if: startsWith(github.ref, 'refs/tags/')
+    needs:
+      - test
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions/setup-node@v3
+        with:
+          node-version: '12'
+          registry-url: https://registry.npmjs.org/
+      - run: npm ci
+      - run: npm publish || true
+        env:
+          NODE_AUTH_TOKEN: ${{ secrets.NPM_TOKEN }}
