+++ .gitignore
@@ -1,4 +1,5 @@
 .DS_Store
 node_modules
-public/
-dist/
+*.js
+*.d.ts
+*.map
