+++ example/public/locales/de/common.json
@@ -1,0 +1,4 @@
+{
+  "Hello": "Hallo",
+  "World": "Welt"
+}
