+++ example/public/locales/en/common.json
@@ -1,0 +1,4 @@
+{
+  "Hello": "Hello",
+  "World": "World"
+}
