+++ tsconfig.json
@@ -1,32 +1,30 @@
 {
-	"compilerOptions": {
-	  "strict": true,
-	  "target": "es6",
-	  "module": "commonjs",
-	  "moduleResolution": "node",
-	  "lib": [
-		"esnext",
-		"dom"
-	  ],
-	  "jsx": "react",
-	  "allowJs": true,
-	  "experimentalDecorators": true,
-	  "emitDecoratorMetadata": true,
-	  "esModuleInterop": true,
-	  "noImplicitAny": true,
-	  "strictPropertyInitialization": false,
-	  "strictNullChecks": false,
-	  "resolveJsonModule": true,
-	  "baseUrl": ".",
-	  "rootDir": "src",
-	  "outDir": "dist",
-	  "sourceMap": true,
-	  "declaration": true
-	},
-	"include": [
-	  "src/**/*"
-	],
-	"exclude": [
-	  "**/*.test.ts"
-	]
-  }
+  "compilerOptions": {
+    "strict": true,
+    "target": "es6",
+    "module": "commonjs",
+    "moduleResolution": "node",
+    "lib": [
+      "esnext",
+      "dom"
+    ],
+    "experimentalDecorators": true,
+    "emitDecoratorMetadata": true,
+    "esModuleInterop": true,
+    "noImplicitAny": true,
+    "strictPropertyInitialization": false,
+    "strictNullChecks": false,
+    "resolveJsonModule": true,
+    "baseUrl": ".",
+    "rootDir": "src",
+    "outDir": ".",
+    "sourceMap": true,
+    "declaration": true
+  },
+  "include": [
+    "src/**/*"
+  ],
+  "exclude": [
+    "**/*.test.ts"
+  ]
+}
