--- webpack.config.js
@@ -1,62 +1,0 @@
-const path = require('path');
-const sveltePreprocess = require('svelte-preprocess');
-
-const mode = process.env.NODE_ENV || 'development';
-const prod = mode === 'production';
-
-module.exports = {
-	entry: './src/index.ts',
-	mode: 'production',
-	resolve: {
-		alias: {
-			svelte: path.dirname(require.resolve('svelte/package.json'))
-		},
-		extensions: ['.mjs', '.js', '.ts', '.svelte'],
-		mainFields: ['svelte', 'browser', 'module', 'main']
-	},
-	output: {
-		path: path.join(__dirname, 'dist'),
-		filename: 'bundle.js',
-		publicPath: '',
-		libraryTarget: 'amd',
-	},
-	module: {
-		rules: [
-			{
-				test: /\.ts$/,
-				loader: 'ts-loader',
-				exclude: /node_modules/
-			},
-			{
-				test: /\.svelte$/,
-				use: {
-					loader: 'svelte-loader',
-					options: {
-						compilerOptions: {
-							dev: !prod
-						},
-						emitCss: prod,
-						hotReload: !prod,
-						preprocess: sveltePreprocess({ sourceMap: !prod })
-					}
-				}
-			},
-			{
-				// required to prevent errors from Svelte on Webpack 5+
-				test: /node_modules\/svelte\/.*\.mjs$/,
-				resolve: {
-					fullySpecified: false
-				}
-			}
-		]
-	},
-	plugins: [],
-	devtool: prod ? false : 'source-map',
-	devServer: {
-		headers: {
-			'Access-Control-Allow-Origin': '*',
-			'Access-Control-Allow-Methods': 'GET, POST, PUT, DELETE, PATCH, OPTIONS',
-			'Access-Control-Allow-Headers': 'X-Requested-With, content-type, Authorization',
-		},
-	}
-};
